Refactor naming validators
